More debugging msg removed + code update.

+++ src/Bpp/Phyl/Model/Protein/DSO78.cpp
@@ -80,8 +80,8 @@ void DSO78::setFreqFromData(const SequenceContainer& data)
   std::map<int, double> freqs;
   SequenceContainerTools::getFrequencies(data, freqs);
   double t = 0;
-  for (unsigned int i = 0; i < size_; i++) t += freqs[i];
-  for (unsigned int i = 0; i < size_; i++) freq_[i] = freqs[i] / t;
+  for (size_t i = 0; i < size_; i++) t += freqs[static_cast<int>(i)];
+  for (size_t i = 0; i < size_; i++) freq_[i] = freqs[static_cast<int>(i)] / t;
   freqSet_->setFrequencies(freq_);
   //Update parameters and re-compute generator and eigen values:
   matchParametersValues(freqSet_->getParameters());
